refactor: modify package structure of the resolver to be more feature sliced
